Add files via upload
Part of the Group Final Project Report (mostly work done by Feifei - we collaborated on data analysis / feedback for each other's reports). I'm not sure if I should be including this in my own GitHub repository specifically (vs the group dropbox we both submitted to)? I'm adding it to this repository in case it's needed for reference, since it was part of our group submission. 

Signed-off-by: izzyvmika <izzy.mika@my.metrostate.edu>